# internal/swagger/api_v1.0.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "GTB Swagger",
      "description": "GTB server API",
      "contact": {
        "email": "developer@energie-ip.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "/{basePath}",
        "variables": {
          "basePath": {
            "default": "v1.0"
          }
        }
      }
    ],
    "paths": {
...
      "/setup/blind/{mac}": {
        "get": {
          "tags": [
            "setup"
          ],
          "summary": "getBlindSetup",
          "description": "Return Blind Setup configuration",
          "operationId": "GetBlindSetup",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Blind Mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BlindSetup"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        },
        "delete": {
          "tags": [
            "setup"
          ],
          "summary": "removeBlindSetup",
          "description": "Remove Blind Setup configuration. It means the driver Blind will be blacklisted",
          "operationId": "RemoveBlindSetup",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Blind Mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
      "/setup/blind": {
        "post": {
          "tags": [
            "setup"
          ],
          "summary": "setBlindSetup",
          "description": "Sensor Blind configuration",
          "operationId": "SetBlindSetup",
          "parameters": [],
          "requestBody": {
            "description": "Blind config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlindSetup"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BlindSetup"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/config/blind": {
        "post": {
          "tags": [
            "config"
          ],
          "summary": "setBlindConfig",
          "description": "Change Blind runtime configuration",
          "operationId": "SetBlindConfig",
          "parameters": [],
          "requestBody": {
            "description": "Blind runtime config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlindConfig"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation, Configuration read from database",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/command/blind": {
        "post": {
          "tags": [
            "command"
          ],
          "summary": "sendBlindCommand",
          "description": "Send Blind command",
          "operationId": "SendBlindCommand",
          "parameters": [],
          "requestBody": {
            "description": "Blind Command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlindCommand"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/status/blind/{mac}": {
        "get": {
          "tags": [
            "status"
          ],
          "summary": "getBlindStatus",
          "description": "Return the given Blind status",
          "operationId": "GetBlindStatus",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Blind Mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BlindStatus"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "ActionProfile": {
          "title": "ActionProfile",
          "required": [
            "action"
          ],
          "type": "object",
          "properties": {
            "action": {
              "type": "integer",
              "description": "Action on the Blind: 0:stop, 1 Openning, 2: Closing",
              "format": "int32"
            },
            "duration": {
              "type": "integer",
              "description": "Action duration in milliseconds",
              "format": "int32"
            }
          }
        },
        "BlindProfile": {
          "title": "BlindProfile",
          "required": [
            "up",
            "down",
            "stop",
            "windowOpen",
            "slatDuration"
          ],
          "type": "object",
          "properties": {
            "up": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ActionProfile"
              },
              "description": "Action sequences"
            },
            "down": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ActionProfile"
              },
              "description": "Action sequences"
            },
            "stop": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ActionProfile"
              },
              "description": "Action sequences"
            },
            "windowOpen": {
              "type": "boolean",
              "description": "When the window is open, is the driver still up and running?"
            },
            "slatDuration": {
              "type": "string",
              "description": "Timing for the slat configuration in Milliseconds"
            }
          }
        },
        "BlindSetup": {
          "title": "BlindSetup",
          "required": [
            "mac",
            "profile"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Blind mac address"
            },
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "isBleEnabled": {
              "type": "boolean",
              "description": "bluetooth activation"
            },
            "profile": {
              "$ref": "#/components/schemas/BlindProfile"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "dumpFrequency": {
              "type": "integer",
              "description": "Dump frequency by default: 1s",
              "format": "int32"
            }
          }
        },
        "BlindConfig": {
          "title": "BlindConfig",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Blind mac address"
            },
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "isBleEnabled": {
              "type": "boolean",
              "description": "bluetooth activation"
            },
            "profile": {
              "$ref": "#/components/schemas/BlindProfile"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "dumpFrequency": {
              "type": "integer",
              "description": "Dump frequency by default: 1s",
              "format": "int32"
            },
            "isConfigured": {
              "type": "boolean",
              "description": "Is the Blind configured"
            }
          }
        },
        "BlindCommand": {
          "title": "BlindCommand",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Blind mac address"
            },
            "blind1": {
              "type": "integer",
              "description": "blind1 0: stop, 1: Openning, 2: Closing",
              "format": "int32"
            },
            "blind2": {
              "type": "integer",
              "description": "blind2 0: stop, 1: Openning, 2: Closing",
              "format": "int32"
            },
            "slat1": {
              "type": "integer",
              "description": "Slat orientation",
              "format": "int32"
            },
            "slat2": {
              "type": "integer",
              "description": "Slat orientation",
              "format": "int32"
            }
          }
        },
...
        "BlindStatus": {
          "title": "BlindStatus",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "LED mac address"
            },
            "ip": {
              "type": "string",
              "description": "Led IP address"
            },
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "protocol": {
              "type": "string",
              "description": "Sensor communication protocol"
            },
            "switchMac": {
              "type": "string",
              "description": "Associate switch mac address"
            },
            "isConfigured": {
              "type": "boolean",
              "description": "Is the sensor configured"
            },
            "softwareVersion": {
              "type": "number",
              "description": "Sensor Software version"
            },
            "hardwareVersion": {
              "type": "string",
              "description": "Sensor Hardware version"
            },
            "isBleEnabled": {
              "type": "boolean",
              "description": "Is the bluetooth activate"
            },
            "error": {
              "type": "integer",
              "description": "Error code value: 0 means no error",
              "format": "int32"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "daisyChainEnabled": {
              "type": "boolean",
              "description": "Daisy chained enabled"
            },
            "daisyChainPos": {
              "type": "integer",
              "description": "Daisy chained position",
              "format": "int32"
            },
            "voltageInput": {
              "type": "integer",
              "description": "Voltage input",
              "format": "int32"
            },
            "dumpFrequency": {
              "type": "integer",
              "description": "Dump frequency by default: 1s",
              "format": "int32"
            },
            "windowStatus1": {
              "type": "boolean",
              "description": "Is the window 1 open"
            },
            "windowStatus2": {
              "type": "boolean",
              "description": "Is the window 2 open"
            }
          }
        },
...
        "Error": {
          "title": "Error",
          "required": [
            "code",
            "message"
          ],
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "message": {
              "type": "string"
            }
          }
        },
...
      },
      "securitySchemes": {
        "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT",
            "description": "For accessing the API a valid JWT token must be passed in all the queries in the Authorization header. Authorization header : Bearer xxxxxx.yyyyyyy.zzzzzz"
        }
      }
    },
    "security": [
      {
        "Authorization": []
      }
    ],
...
  }